10-10d | Organize tests directory
