+++ src/applications/ivc-champva/10-10d-extended/tests/e2e/10-10d-extended.cypress.spec.js
@@ -2,11 +2,8 @@ import path from 'path';
 import testForm from 'platform/testing/e2e/cypress/support/form-tester';
 import { createTestConfig } from 'platform/testing/e2e/cypress/support/form-tester/utilities';
 import mockUser from './fixtures/mocks/user.json';
-import formConfig from '../config/form';
-import manifest from '../manifest.json';
-// TODO: uncomment this mock feature toggle when we get a functioning E2E test
-// up and running:
-// import mockFeatureToggles from './e2e/fixtures/mocks/featureToggles.json';
+import formConfig from '../../config/form';
+import manifest from '../../manifest.json';
 
 const testConfig = createTestConfig(
   {
