+++ src/applications/ivc-champva/10-10d-extended/tests/e2e/fixtures/data/maximal-test.json
@@ -1,0 +1,105 @@
+{
+  "data": {
+    "view:hasApplicants": true,
+    "applicants": [
+      {
+        "applicantRelationshipOrigin": {
+          "relationshipToVeteran": "blood"
+        },
+        "applicantEmailAddress": "johnny@alvin.gov",
+        "applicantAddress": {
+          "country": "USA",
+          "street": "456 Circle Street",
+          "city": "Clinton",
+          "state": "AS",
+          "postalCode": "56790"
+        },
+        "applicantSSN": "345345345",
+        "applicantPhone": "555-555-1234",
+        "applicantGender": {
+          "gender": "male"
+        },
+        "applicantRelationshipToSponsor": {
+          "relationshipToVeteran": "child"
+        },
+        "applicantDependentStatus": {
+          "status": "enrolled"
+        },
+        "applicantEnrolledInOHI": false,
+        "applicantDob": "2003-01-04",
+        "applicantName": {
+          "first": "Johnny",
+          "middle": "T",
+          "last": "Alvin",
+          "suffix": "Jr."
+        }
+      }
+    ],
+    "healthInsurance": [
+      {
+        "insuranceType": "medigap",
+        "medigapPlan": "K",
+        "provider": "Blue Cross Blue Shield",
+        "effectiveDate": "2024-10-01",
+        "expirationDate": "2024-10-02",
+        "throughEmployer": true,
+        "healthcareParticipants": {
+          "28e7f1064a74": true
+        },
+        "insuranceCard": [],
+        "additionalComments": "Lorem"
+      }
+    ],
+    "medicare": [
+      {
+        "medicareParticipant": "28e7f1064a74",
+        "medicarePlanType": "c",
+        "medicarePartAEffectiveDate": "2023-01-15",
+        "medicarePartBEffectiveDate": "2023-01-15",
+        "medicarePartAPartBFrontCard": [],
+        "medicarePartAPartBBackCard": [],
+        "medicarePartCCarrier": "Advantage Health Solutions",
+        "medicarePartCEffectiveDate": "2023-02-01",
+        "hasPharmacyBenefits": true,
+        "medicarePartCFrontCard": [],
+        "medicarePartCBackCard": [],
+        "hasMedicarePartD": true,
+        "medicarePartDCarrier": "PharmaCare Plus",
+        "medicarePartDEffectiveDate": "2023-02-01",
+        "medicarePartDFrontCard": [],
+        "medicarePartDBackCard": []
+      }
+    ],
+    "sponsorSsn": "123123123",
+    "sponsorName": {
+      "first": "Joe",
+      "last": "Johnson"
+    },
+    "certifierRole": "other",
+    "certifierRelationship": {
+      "relationshipToVeteran": {
+        "spouse": true,
+        "child": true,
+        "thirdParty": true
+      }
+    },
+    "certifierPhone": "1231231234",
+    "certifierEmail": "certifier@email.gov",
+    "certifierAddress": {
+      "country": "USA",
+      "street": "123 Certifier Street",
+      "city": "Citytown",
+      "state": "AL",
+      "postalCode": "12312"
+    },
+    "certifierName": {
+      "first": "Certifier",
+      "last": "Jones"
+    },
+    "sponsorDob": "1958-01-01",
+    "sponsorIsDeceased": true,
+    "sponsorDOD": "2001-01-01",
+    "sponsorDeathConditions": false,
+    "statementOfTruthSignature": "Certifier Jones"
+  }
+}
