+++ src/applications/ivc-champva/10-10d-extended/tests/e2e/fixtures/mocks/maintenance-windows.json
@@ -1,0 +1,3 @@
+{
+  "data": []
+}
