+++ src/applications/ivc-champva/10-10d-extended/tests/e2e/fixtures/mocks/vamc-ehr.json
@@ -1,0 +1,8 @@
+{
+  "data": {
+    "nodeQuery": {
+      "count": 0,
+      "entities": []
+    }
+  }
+}
