+++ src/applications/ivc-champva/10-10d-extended/tests/unit/containers/ConfirmationPage.unit.spec.jsx
@@ -4,9 +4,9 @@ import { createStore } from 'redux';
 import { Provider } from 'react-redux';
 import { cleanup, render } from '@testing-library/react';
 import { createInitialState } from '@department-of-veterans-affairs/platform-forms-system/state/helpers';
-import formConfig from '../../config/form';
-import ConfirmationPage from '../../containers/ConfirmationPage';
-import mockData from '../fixtures/data/test-data.json';
+import formConfig from '../../../config/form';
+import ConfirmationPage from '../../../containers/ConfirmationPage';
+import mockData from '../../e2e/fixtures/data/maximal-test.json';
 
 const mockStore = state => createStore(() => state);
