+++ src/applications/ivc-champva/10-10d-extended/tests/unit/containers/IntroductionPage.unit.spec.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { render } from '@testing-library/react';
 import { expect } from 'chai';
-import formConfig from '../../config/form';
-import IntroductionPage from '../../containers/IntroductionPage';
+import formConfig from '../../../config/form';
+import IntroductionPage from '../../../containers/IntroductionPage';
 
 const props = {
   route: {
